supplement consent code models
