+++ luce_vm/scripts/simulation/mimic_consent_codes.py
@@ -6,112 +6,93 @@ class ConsentCodeSimulator:
         self.primary_categories = [
             "NRES", "GRU(CC)", "HMB(CC)", "DS-[XX](CC)", "POA"
         ]
-
         self.secondary_categories = ["RS-[XX]", "RUO", "NMDS", "GSO"]
         self.requirements = [
-            "NPU", "PUB", "COL-[XX]", "IRB", "GS-[XX]", "MOR-[XX]", "TS-[XX]",
-            "US", "PS", "IS"
+            "NPU", "PUB", "IRB", "GS-[XX]", "MOR-[XX]", "TS-[XX]", "US", "PS",
+            "IS"
         ]
 
-        # To define any logical constraints or dependencies among codes
-        self.constraints = {
-            # If "NRES" is chosen, exclude all secondary categories and requirements,
-            # as "No restrictions" implies complete openness with no additional conditions.
-            "NRES": {
-                "exclude": self.secondary_categories + self.requirements
-            },
+    def get_valid_consent_codes(self):
 
-            # If "RUO" (Research use only) is chosen as a secondary category, exclude any
-            # contradictory categories or requirements that imply non-research use or commercial use.
-            "RUO": {
-                "exclude": ["CommercialUseRequirement"]
-            },  # Replace with actual codes implying commercial use.
+        valid_codes = {}
+        is_valid = False
+        while not is_valid:
+            consent_code = self.generate_consent_code()
+            is_valid = self.validate_consent_code(consent_code['Primary'],
+                                                  consent_code['Secondary'],
+                                                  consent_code['Requirements'])
+        return consent_code
 
-            # "NPU" (Not-for-profit use only) would logically exclude any requirements
-            # or conditions related to profit-oriented or commercial use.
-            "NPU": {
-                "exclude": ["CommercialUseRequirement"]
-            },  # Replace with actual codes implying commercial use.
+    def random_num_secondary_categories(self):
+        return random.randint(0, len(self.secondary_categories))
 
-            # "PUB" (Publication required) may have conditional complementarity with "MOR-[XX]"
-            # (Publication moratorium/embargo). They can be selected together if the moratorium
-            # period is respected, meaning results cannot be published until after the specified date.
-            "PUB": {
-                "complement": ["MOR-[XX]"]
-            },
+    def random_num_requirements(self):
+        return random.randint(0, len(self.requirements))
 
-            # If a disease-specific category "DS-[XX](CC)" is chosen, exclude any contradictory
-            # secondary categories or requirements that imply broader or different disease applicability.
-            "DS-[XX](CC)": {
-                "exclude": ["ContradictoryCategory"]
-            },  # Replace with actual contradictory categories.
+    def select_primary_category(self):
+        return random.choice(self.primary_categories)
 
-            # More concrete constraints based on specific definitions, implications,
-            # and interactions of each consent code should be added here.
-        }
+    def select_secondary_categories(self, num_categories):
+        return random.sample(self.secondary_categories, num_categories)
 
-    def generate_random_combination(self):
-        primary = random.choice(self.primary_categories)
+    def select_requirements(self, num_requirements):
+        return random.sample(self.requirements, num_requirements)
 
-        possible_secondaries = [
-            sec for sec in self.secondary_categories
-            if not self.check_constraint(primary, sec)
-        ]
-        secondaries = random.sample(possible_secondaries,
-                                    k=random.randint(
-                                        0, len(possible_secondaries)))
-
-        possible_requirements = [
-            req for req in self.requirements
-            if not any(self.check_constraint(s, req) for s in secondaries)
-        ]
-        requirements = random.sample(possible_requirements,
-                                     k=random.randint(
-                                         0, len(possible_requirements)))
+    def generate_consent_code(self):
+        primary = self.select_primary_category()
+        num_secondary = self.random_num_secondary_categories()
+        num_requirements = self.random_num_requirements()
+        secondaries = self.select_secondary_categories(num_secondary)
+        requirements = self.select_requirements(num_requirements)
 
         return {
-            "primary": primary,
-            "secondary": secondaries,
-            "requirements": requirements,
+            'Primary': primary,
+            'Secondary': secondaries,
+            'Requirements': requirements
         }
 
-    def check_constraint(self, primary, secondary_or_requirement):
-        if primary in self.constraints:
-            return secondary_or_requirement in self.constraints[primary].get(
-                'exclude', [])
-        return False
+    def validate_consent_code(self, primary, secondaries, requirements):
+        # No Restrictions (NRES) Rule: When "No restrictions" is the primary category, no secondary categories or requirements should be allowed, as it defeats the purpose of "no restrictions."
+        if primary == 'NRES' and (len(secondaries) > 0
+                                  or len(requirements) > 0):
+            return False
 
-    def validate_combination(self, combination):
-        primary = combination['primary']
+        # Disease-Specific (DS) Rule: When selecting a disease-specific primary category, there should be at least one requirement, as disease-specific data usually have stringent constraints.
+        if primary.startswith('DS-') and len(requirements) == 0:
+            return False
 
-        for secondary in combination['secondary']:
-            if self.check_constraint(primary, secondary):
-                return False
+        # Research Use Only (RUO) Rule: If "Research use only" is chosen as a secondary category, "General research use and clinical care" should not be the primary category, since clinical care is not just for research.
+        if primary == 'GRU(CC)' and 'RUO' in secondaries:
+            return False
 
-        for requirement in combination['requirements']:
-            if any(
-                    self.check_constraint(s, requirement)
-                    for s in combination['secondary']):
-                return False
+        # Genetic Studies Only (GSO) Rule: If this is a secondary category, "Publication required" should be a requirement to ensure ethical usage of genetic data.
+        if 'GSO' in secondaries and 'PUB' not in requirements:
+            return False
 
-        return True
+        # Not-for-profit use only (NPU) Rule: If this is a requirement, "General research use and clinical care" should not be the primary category, as clinical care might involve for-profit healthcare providers.
+        if primary == 'GRU(CC)' and 'NPU' in requirements:
+            return False
 
-    def simulate(self):
-        valid_combinations = []
-        invalid_combinations = []
+        # ... (add more rules as necessary)
 
-        for _ in range(10000):  # Replace with desired number of simulations
-            combination = self.generate_random_combination()
-            if self.validate_combination(combination):
-                valid_combinations.append(combination)
-            else:
-                invalid_combinations.append(combination)
-
-        return valid_combinations, invalid_combinations
+        return True  # Return True if all validation rules pass
 
 
+# Create an instance of this class
 simulator = ConsentCodeSimulator()
-valid_combinations, invalid_combinations = simulator.simulate()
 
-print("Number of valid combinations: ", len(valid_combinations))
-print("Number of invalid combinations: ", len(invalid_combinations))
+# Generate a consent code with 2 secondary categories and 3 requirements
+consent_code = simulator.generate_consent_code()
+
+print(f"Generated Consent Code: {consent_code}")
+
+# Validate the consent code
+is_valid = simulator.validate_consent_code(consent_code['Primary'],
+                                           consent_code['Secondary'],
+                                           consent_code['Requirements'])
+
+print(f"Is Valid: {is_valid}")
+
+# Generate a valid consent code
+valid_consent_code = simulator.get_valid_consent_codes()
+print(f"Valid Consent Code: {valid_consent_code}")
